Fixed and implemented spells and pickups
As a designer, I want the spells to be randomly chosen from a pool when they spawn in, so that the players can enjoy and strategize during the matches

As a designer, I want the players to only have one spell slot each, so that the players don't hog all the spells and need to strategize on how to use the spells they've picked up.

As a designer, I want spells to be able to be cast, so that I can test their effects.